Moved template function to .h
