+++ src/software/embedded/BUILD
@@ -44,7 +44,6 @@ cc_library(
 
 cc_library(
     name = "spi_utils",
-    srcs = ["spi_utils.cpp"],
     hdrs = ["spi_utils.h"],
     deps = [
         "//software/logger",
